refactor polyutils into multiple files
